Integration tests working for all but datadiscovery and trials
