+++ etl/transform/utils.py
@@ -1,5 +1,12 @@
 import base64
+import glob
+import gzip
+import json
+import os
+import re
+import shutil
 import urllib.parse
+from xml.sax import saxutils as su
 
 import rfc3987
 
@@ -53,3 +60,82 @@ def get_generated_uri_from_str(source: dict, entity: str, data: str, do_base64 =
     if do_base64:
         data_uri = base64.b64encode(data_uri.encode('utf-8')).decode('utf-8')
     return data_uri
+
+
+def is_checkpoint(n):
+    return n > 0 and n % 10000 == 0
+
+
+def save_json(source_dir, json_dict, logger):
+    logger.debug("Saving documents to json files...")
+    saved_documents = 0
+    for type, documents in json_dict.items():
+        file_number = 1
+        saved_documents = 0
+        documents_list = documents.values()
+        while saved_documents < len(documents_list):
+            with open(source_dir + "/" + type + '-' + str(file_number) + '.json', 'w') as f:
+                json.dump(list(documents_list)[saved_documents:file_number * 10000], f, ensure_ascii=False)
+            with open(source_dir + "/" + type + '-' + str(file_number) + '.json', 'rb') as f:
+                with gzip.open(source_dir + "/" + type + '-' + str(file_number) + '.json.gz', 'wb') as f_out:
+                    shutil.copyfileobj(f, f_out)
+            os.remove(source_dir + "/" + type + '-' + str(file_number) + '.json')
+            file_number += 1
+            saved_documents += 10000
+            logger.debug(f"checkpoint: {saved_documents} documents saved")
+    logger.debug(f"Total of {saved_documents} documents saved in json files.")
+
+
+def remove_html_tags(text):
+    """
+    Remove html tags from a string
+    """
+    extra_char = {
+        '&apos;': '',
+        '&quot;': '',
+        '&amp;': ''
+    }
+    # unescap HTML tags
+    text = su.unescape(text, extra_char)
+    clean = re.compile('<.*?>')
+    return re.sub(clean, '', text)
+
+
+def json_to_jsonl(source_json_dir):
+    """
+    Conversion from JSON to JSONL (http://jsonlines.org/) for EVA dump
+    :param source_json_dir: the json files directory
+    """
+    json_files = glob.glob(source_json_dir + "/*.json")
+    for json_file in json_files:
+        # read the file
+        try:
+            with open(json_file) as old_json_file:
+                data = json.load(old_json_file)
+        except json.decoder.JSONDecodeError:
+            print("INFO: The file '{}' is already flattened. Removing HTML tags if any ..".format(json_file))
+            continue
+        # write the new one (overriding the old json)
+        with open(json_file, 'w') as new_json_file:
+            for entry in data:
+                json.dump(entry, new_json_file)
+                new_json_file.write('\n')
+
+
+def rm_tags(source_json_dir):
+    json_files = glob.glob(source_json_dir + "/*.json")
+    for json_file in json_files:
+        new_json_list = []
+        with open(json_file) as old_json_file:
+            json_list = list(old_json_file)
+        for json_str in json_list:
+            line = json.loads(json_str)
+            if "studyDescription" in line:
+                # remove escaped html
+                line["studyDescription"] = remove_html_tags(line["studyDescription"])
+            new_json_list.append(line)
+
+        with open(json_file, 'w') as new_json_file:
+            for entry in new_json_list:
+                json.dump(entry, new_json_file)
+                new_json_file.write('\n')
